Update patch set 1
Patch Set 1:

Many of the special keys are Samsung specific and do not work in stock Android so those have been disabled. However the RECENTAPPS code has been changed to APP_SWITCH which is supported.

Also with the correct layout being used the BACK key will be a real BACK and not ESC as it is being mapped if it falls back to Generic.kl (which it does if it finds one unsupported keycode in the layout.

But you won't notice much different because of the many Samsung specific keys not being supported in CM/AOSP. However there is no point including these layouts if they cannot be loaded by Android so therefor I've removed the unsupported keycodes and change RECENTAPPS into APP_SWITCH.

I short, I think the only difference you will notice is that BACK works better and that the APP_SWITCH key starts working.

Patch-set: 1